+++ CHANGELOG.md
@@ -4,8 +4,9 @@
 - Lib: Replace some unsafe blocks by using `bytemuck` instead
 - Lib: Redesign the API for SPH kernel functions
 - Lib: Add module `sph_interpolation` for the interpolation of normals, scalar fields and vector fields to arbitrary points using SPH based interpolation
-- CLI: Add command line switch `--sph-normals` (enabled by default) to control whether to use SPH interpolation to compute normals when `--output-normals` is on
+- CLI: Add command line switch `--sph-normals=on/off` (enabled by default) to control whether to use SPH interpolation to compute normals when `--output-normals` is on
 - CLI: Actually support writing surfaces meshes (including normals) to OBJ files (https://github.com/w1th0utnam3/splashsurf/pull/29)
+- CLI: Support for specifying names of point attribute fields via the argument `--attributes=...` that should be interpolated to the reconstructed surface. Currently it is only possible to load these fields from VTK files and the interpolated values can also only be written if the output file format is VTK.
 
 ## Version 0.7.0
 
@@ -13,7 +14,7 @@
  - Lib: Update to `nalgebra` 0.27. with const generics and removed usage of `nalgebra` types that are deprecated
  - Lib: Add a "naive" `O(n^2)` neighborhood search for testing purposes
  - Lib: Rename other neighborhood search functions to `neighborhood_search_spatial_hashing` and `neighborhood_search_spatial_hashing_parallel`
- - CLI: Add a command line option `--output-normals` to the `reconstruct` sub-command to write mesh normals to the output file. Note that currently the normals are only computed using an area weighted average of triangle normals.
+ - CLI: Add a command line option `--output-normals=on/off` to the `reconstruct` sub-command to write mesh normals to the output file. Note that currently the normals are only computed using an area weighted average of triangle normals.
  - CLI: Change the `convert` sub-command to allow both conversion of particle files and conversion of mesh formats. Instead of using the `--input` option you now have to either use the `--particles` or `--mesh` option to specify the input file depending on whether you want to convert a particle or mesh file. For particles `VTK, BGEO, PLY, XYZ, JSON -> VTK, PLY` is supported. For meshes only `VTK, PLY -> OBJ` is supported.
  - CLI: Add JSON files as supported particle input file
  - CLI: Change the command line option to specify a sequence of input files from `--input-sequence-pattern` to `--input-sequence`